refactor: share terminal agent status parser
